web-fe/requirements.txt
gradio==4.26.0 
langchain==0.1.11
websocket-client==0.53.0
